redesign/website/requirements.txt
Flask==3.0.0
google-cloud-firestore==2.19.0